requirements.txt
annotated-types==0.7.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
confection==0.1.5
cryptography==44.0.3
cymem==2.0.11
dateparser==1.2.2
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.0
filelock==3.18.0
idna==3.10
iniconfig==2.1.0
jellyfish==1.2.0
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.1
numpy==2.0.2
packaging==25.0
pandas==2.3.0
pdfminer.six==20250506
pdfplumber==0.11.7
phonenumbers==8.13.55
pillow==11.3.0
pluggy==1.6.0
preshed==3.0.10
presidio_analyzer==2.2.358
presidio_anonymizer==2.2.358
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.1
pypdfium2==4.30.1
pytesseract==0.3.13
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
skillNer==1.0.3
smart_open==7.3.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.6
threadpoolctl==3.6.0
tldextract==5.3.0
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
transformers==4.41.2
torch==2.3.1
sentencepiece==0.2.0
marker-pdf==0.2.3
